Back-end/scanner/xss_scanner.py
from algosdk.v2client import algod
# from models.vulnerability import Vulnerability, ScanResult
from datetime import datetime

# Mocking the necessary functions and classes for the example
class MockAlgodClient:
    def application_info(self, app_id):
        return {
            'params': {
                'approval-program': 'mock_teal_code',
                'global-state': [{'key': 'example_key', 'value': 'example_value'}]
            }
        }

def fetch_vulndb_data(teal_code):
    # Simulating a successful fetch with no vulnerabilities
    return []

def scan_app_state(app_state):
    # Simulating a successful scan with no vulnerabilities
    return []

# Mocking the Vulnerability and ScanResult classes
class MockVulnerability:
    def __init__(self, risk_level, recommendations):
        self.risk_level = risk_level
        self.recommendations = recommendations

class MockScanResult:
    def __init__(self, app_id, vulnerabilities, risk_level, recommendations, scan_timestamp):
        self.app_id = app_id
        self.vulnerabilities = vulnerabilities
        self.risk_level = risk_level
        self.recommendations = recommendations
        self.scan_timestamp = scan_timestamp

# Mocking the XSSPatternMatcher class
class MockXSSPatternMatcher:
    pass

# Implementing the AlgorandXSSScanner using mocked classes
class AlgorandXSSScanner:
    def __init__(self, algod_address: str, algod_token: str):
        self.algod_client = MockAlgodClient()
        self.pattern_matcher = MockXSSPatternMatcher()

    def scan_contract(self, app_id: int) -> MockScanResult:
        vulnerabilities = []

        # Get contract info
        contract_info = self._get_contract_info(app_id)
        if contract_info:
            # Scan TEAL code
            teal_vulnerabilities = self._scan_teal_code(
                contract_info.get('params', {}).get('approval-program', '')
            )
            vulnerabilities.extend(teal_vulnerabilities)

            # Scan application state
            state_vulnerabilities = self._scan_application_state(self._get_application_state(app_id))
            vulnerabilities.extend(state_vulnerabilities)

        # Determine overall risk level
        risk_level = self._calculate_risk_level(vulnerabilities)

        # Generate recommendations
        recommendations = self._generate_recommendations(vulnerabilities)

        return MockScanResult(
            app_id=app_id,
            vulnerabilities=vulnerabilities,
            risk_level=risk_level,
            recommendations=recommendations,
            scan_timestamp=datetime.now().isoformat()
        )

    def _get_contract_info(self, app_id: int):
        return self.algod_client.application_info(app_id)

    def _get_application_state(self, app_id: int):
        response = self.algod_client.application_info(app_id)
        if 'params' in response and 'global-state' in response['params']:
            global_state = response['params']['global-state']
            global_state_dict = {item['key']: item['value'] for item in global_state}
            return global_state_dict
        else:
            return None

    def _scan_teal_code(self, contract_info):
        return fetch_vulndb_data(contract_info)

    def _scan_application_state(self, app_state):
        return scan_app_state(app_state)

    def _calculate_risk_level(self, vulnerabilities):
        if vulnerabilities:
            return 'Low'  # Assuming no vulnerabilities found
        return 'None'

    def _generate_recommendations(self, vulnerabilities):
        return 'No recommendations needed.' if not vulnerabilities else 'Take necessary actions.'

# Example usage of the scanner
if __name__ == "__main__":
    algod_address = "http://localhost:4001"  # Mock address
    algod_token = "your_algod_token"  # Mock token
    app_id = 123456  # Example application ID

    scanner = AlgorandXSSScanner(algod_address, algod_token)
    result = scanner.scan_contract(app_id)

    print(f"Scan Result for App ID {result.app_id}:")
    print(f"Vulnerabilities: {result.vulnerabilities}")
    print(f"Risk Level: {result.risk_level}")
    print(f"Recommendations: {result.recommendations}")
    print(f"Scan Timestamp: {result.scan_timestamp}")
